Added the ability track mouse input along with keyboard input
Possible outputs are now set to be a special permutation of any given list. This is only because technically you can press any permutation of buttons and still have the vehicle do something. In this example, many inputs will not be used excessively. Later we will define a threshold for how many times the input must appear before we train to it.