...over a year later, finally fixing/updating the GitHub
The repo is in a completely broken and very unusable state at the
moment. I'll be fixing it in the coming days/weeks, hopefully to
once again create a self driving car -- except this time better!

Yes, `grabscreen.py` is entirely commented out. It was originally
designed to only work with Windows. As you could imagine, it doesn't
play nicely with Linux. Will be redoing it soon™.